+++ examples/connext_dds/network_capture/01_shared_memory_and_udp/cs/NetworkCapture.idl
@@ -9,6 +9,10 @@
  * not be liable for any incidental or consequential damages arising out of the
  * use or inability to use the software.
  */
-struct NetworkCaptureData {
-    string<128> msg;
+module NetworkCaptureExample {
+
+    struct HelloWorld {
+        @key string<128> message;
+    };
+
 };
